Till now Database is connected and student portal is completed partially

+++ StudentMngmtSystem/src/main/java/com/base/repository/AdminRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.base.entity.*;
 
-public interface AdminRepository extends JpaRepository<AdminEntity, Long>{
+public interface AdminRepository extends JpaRepository<AdminEntity, Integer>{
 	
 	public AdminEntity findByAdminLoginAndAdminPassword(String adminLogin, String adminPassword);
